Moved custom hook into reusable function
